+++ src/components/ui/button/index.ts
@@ -1,1 +1,1 @@
-export * from './button'
+export * from './Button'
